refactor: create simple repository implementation
